Arbitrum regression test
Test fix: introduce delay between putting the vault for auction and liquidating it.
Otherwise, not enough collateral is released and the swap fails

Other changes:
 * reduce bot logging
 * add @hardhat-dapptools to help with failed tx debugging
 * 'make build' now doesn't rebuild yarn.lock. Instead, it tries to use it and
complains if it needs to be updated. Reduces churn in yarn.lock
 * switching to yarn 3 to speed up installation
 * fixed the uniswap router: @uniswap/smart-order-router has a gas estimation bug that prevents it from
working with local hardhat node. Something in my local node_modules structure prevented it
from hitting that bug (it fell back to estimate gas using external service which is bad - not chain
specific and not deterministic). I put in a workaround for the bug, see comments in router.ts and this
for details: https://github.com/Uniswap/smart-order-router/issues/68