CI: regression tests, Arbitrum liquidator
 + CI now includes regression tests
 * uni router: increase the timeout
 + liquidator bot: "VaultsWhiteList" config option. If it's set, the bot ignores all vaults not in this whitelist. It is
meant to be used in **tests only** to speed up regression tests (we only check 1 vault instead of 1000)
 + liquidator bot: --blocks-per-batch option. When the bot retrieves events to discover auctions or vaults, it starts with
a broad query: 'give me all VaultPoured events from block 0 to block X'. For some reason, Infura on Arbitrum can't serve this query
if the range of the blocks is >100,000. --blocks-per-batch is an optional setting that splits the original big query into multiple
smaller subqueries
 * small fix to 'collateral offered' computation: instead of using system clock for determining current time we use latest block timestamp.
This fixes the 'CollateralLimit' test that was mistakingly passing before (for a different reason)